refactor(flows): eliminate shallow pass-through modules
Consolidate 6 files → 4 files in common/ by eliminating thin wrappers:

Eliminated:
- types.ts: Merged AgentRunShared/BaseFlowShared into AgentRunFlowRunner.ts
- runStateSchemas.ts: Moved schemas to respective run flow files

Moved to single consumers:
- NodeExecResult → ToolUseRunFlow.ts (only used there)
- ToolUseRunStateSchema → ToolUseRunFlow.ts (defined where exported)
- ReflectionRunStateSchema → ReflectionRunFlow.ts (defined where exported)

Remaining in common/:
- AgentLifecycle.ts (155 lines - substantial logic)
- AgentRunFlowRunner.ts (now 130 lines with merged types)
- createAgentRunFlow.ts (160 lines - factory + node)
- createFinalizeNode.ts (200 lines - factory + helpers)
- index.ts (barrel for convenience)

Each run flow is now self-contained with its own schema definition.
