+++ src/agent/implementations/BaseAgent.ts
@@ -24,8 +24,8 @@ import type { AgentRoundFinalizedCallback } from '@agent/core/AgentSharedStore';
 import {
   runAgentFlow,
   type AgentRunFlowOptions,
+  type AgentRunShared,
 } from '@agent/implementations/flows/common/AgentRunFlowRunner';
-import type { AgentRunShared } from '@agent/implementations/flows/common/types';
 import type { AgentLifecycle } from '@agent/implementations/flows/common/AgentLifecycle';
 import { AgentExecutionContext } from '@agent/runtime/AgentExecutionContext';
 import { retryCoordinator } from '@agent/runtime/RetryRequestCoordinator';
