+++ src/agent/implementations/flows/common/index.ts
@@ -3,16 +3,14 @@
  *
  * Exports:
  * - AgentLifecycle: Phase/status state machine
- * - AgentRunFlowRunner: Flow execution with hooks
+ * - AgentRunFlowRunner: Flow execution with hooks, AgentRunShared, BaseFlowShared
  * - createAgentRunFlow: Flow factory (includes init node, link wiring)
  * - createStandardFinalizeNode: Standard finalize node factory
- * - types: AgentRunShared, BaseFlowShared
- * - runStateSchemas: Serialization schemas
+ *
+ * Note: Schemas moved to respective run flow files (ToolUseRunFlow, ReflectionRunFlow)
  */
 
 export * from './AgentLifecycle';
 export * from './AgentRunFlowRunner';
 export * from './createAgentRunFlow';
 export * from './createFinalizeNode';
-export * from './runStateSchemas';
-export * from './types';
